refactor: extract code generation tests from cli
